Add files via upload
The objective of this analysis.stock is to forecast the price of an apple in the future. By taking into account the historical stock price values from the Apple dataset to predict the changes in stockprice in Python and successfully predicted stock prices for Apple andStream-lit was used to deploy the model, this analysis will take a closer look at the Apple stock price and forecast its future returns.